pyproject.toml
[project]
name = "anon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pip>=25.0.1",
    "presidio-analyzer[transformers]>=2.2.357",
    "presidio-anonymizer>=2.2.357",
    "protobuf>=6.30.1",
    "python-docx>=1.1.2",
    "sentencepiece>=0.2.0",
    "spacy>=3.8.4",
    "torch>=2.6.0",
    "transformers>=4.49.0",
]
